+++ tests/src/main/java/com/github/klee0kai/test/inject/di/IdentityModule.java
@@ -3,9 +3,9 @@ package com.github.klee0kai.test.inject.di;
 
 import com.github.klee0kai.stone.annotations.module.Module;
 import com.github.klee0kai.stone.annotations.module.Provide;
-import com.github.klee0kai.test.inject.identity.Conscience;
-import com.github.klee0kai.test.inject.identity.Ideology;
-import com.github.klee0kai.test.inject.identity.Knowledge;
+import com.github.klee0kai.test.mowgli.identity.Conscience;
+import com.github.klee0kai.test.mowgli.identity.Ideology;
+import com.github.klee0kai.test.mowgli.identity.Knowledge;
 
 
 @Module
