--- tests/src/main/java/com/github/klee0kai/test/inject/forest/Blood.java
@@ -1,15 +1,0 @@
-package com.github.klee0kai.test.inject.forest;
-
-import java.awt.*;
-import java.util.UUID;
-
-public class Blood {
-
-    public UUID uuid = UUID.randomUUID();
-
-    public Color color = null;
-
-    public Blood(Color color) {
-        this.color = color;
-    }
-}
