--- tests/src/main/java/com/github/klee0kai/test/inject/forest/Earth.java
@@ -1,10 +1,0 @@
-package com.github.klee0kai.test.inject.forest;
-
-import java.util.UUID;
-
-public class Earth {
-
-    public UUID uuid = UUID.randomUUID();
-
-
-}
