--- tests/src/main/java/com/github/klee0kai/test/inject/forest/History.java
@@ -1,10 +1,0 @@
-package com.github.klee0kai.test.inject.forest;
-
-import java.util.UUID;
-
-public class History {
-
-    public UUID uuid = UUID.randomUUID();
-
-
-}
