--- tests/src/main/java/com/github/klee0kai/test/inject/forest/IAnimal.java
@@ -1,7 +1,0 @@
-package com.github.klee0kai.test.inject.forest;
-
-public interface IAnimal {
-
-    void born();
-
-}
