--- tests/src/main/java/com/github/klee0kai/test/inject/identity/Conscience.java
@@ -1,14 +1,0 @@
-package com.github.klee0kai.test.inject.identity;
-
-import java.util.UUID;
-
-public class Conscience {
-
-    public UUID uuid = UUID.randomUUID();
-
-    public boolean isOldConscience() {
-        return false;
-    }
-
-
-}
