--- tests/src/main/java/com/github/klee0kai/test/inject/identity/Knowledge.java
@@ -1,14 +1,0 @@
-package com.github.klee0kai.test.inject.identity;
-
-import java.util.UUID;
-
-public class Knowledge {
-
-    public UUID uuid = UUID.randomUUID();
-
-
-    public boolean isOldKnowledge(){
-        return false;
-    }
-
-}
