--- tests/src/main/java/com/github/klee0kai/test/lifecycle/base/APhoneLifecycle.java
@@ -1,11 +1,0 @@
-package com.github.klee0kai.test.lifecycle.base;
-
-public interface APhoneLifecycle {
-
-    void onBuy();
-
-    void onBroken();
-
-    void onDrown();
-
-}
