--- tests/src/main/java/com/github/klee0kai/test/lifecycle/structure/Battery.java
@@ -1,9 +1,0 @@
-package com.github.klee0kai.test.lifecycle.structure;
-
-import java.util.UUID;
-
-public class Battery {
-
-    public final UUID uuid = UUID.randomUUID();
-
-}
