+++ tests/src/main/java/com/github/klee0kai/test/mowgli/animal/IAnimal.java
@@ -1,0 +1,7 @@
+package com.github.klee0kai.test.mowgli.animal;
+
+public interface IAnimal {
+
+    void born();
+
+}
