+++ tests/src/main/java/com/github/klee0kai/test/mowgli/body/Blood.java
@@ -1,0 +1,15 @@
+package com.github.klee0kai.test.mowgli.body;
+
+import java.awt.*;
+import java.util.UUID;
+
+public class Blood {
+
+    public UUID uuid = UUID.randomUUID();
+
+    public Color color = null;
+
+    public Blood(Color color) {
+        this.color = color;
+    }
+}
