+++ tests/src/main/java/com/github/klee0kai/test/mowgli/identity/Conscience.java
@@ -1,0 +1,14 @@
+package com.github.klee0kai.test.mowgli.identity;
+
+import java.util.UUID;
+
+public class Conscience {
+
+    public UUID uuid = UUID.randomUUID();
+
+    public boolean isOldConscience() {
+        return false;
+    }
+
+
+}
