+++ tests/src/main/java/com/github/klee0kai/test/mowgli/identity/Ideology.java
@@ -1,0 +1,14 @@
+package com.github.klee0kai.test.mowgli.identity;
+
+import java.util.UUID;
+
+public class Ideology {
+
+    public UUID uuid = UUID.randomUUID();
+
+
+    public boolean isFamilyIdeology(){
+        return false;
+    }
+
+}
