+++ tests/src/main/java/com/github/klee0kai/test/mowgli/world/History.java
@@ -1,0 +1,10 @@
+package com.github.klee0kai.test.mowgli.world;
+
+import java.util.UUID;
+
+public class History {
+
+    public UUID uuid = UUID.randomUUID();
+
+
+}
