+++ tests/src/main/java/com/github/klee0kai/test/tech/components/Battery.java
@@ -1,0 +1,9 @@
+package com.github.klee0kai.test.tech.components;
+
+import java.util.UUID;
+
+public class Battery {
+
+    public final UUID uuid = UUID.randomUUID();
+
+}
