+++ tests/src/main/java/com/github/klee0kai/test/tech/components/Ram.java
@@ -1,0 +1,20 @@
+package com.github.klee0kai.test.tech.components;
+
+import com.github.klee0kai.test.lifecycle.di.qualifiers.RamSize;
+
+import java.util.UUID;
+
+public class Ram {
+
+    public final UUID uuid = UUID.randomUUID();
+
+    public final String size;
+
+    public Ram() {
+        size = null;
+    }
+
+    public Ram(RamSize ramSize) {
+        this.size = ramSize.size;
+    }
+}
