+++ tests/src/main/java/com/github/klee0kai/test/tech/phone/base/APhoneLifecycle.java
@@ -1,0 +1,11 @@
+package com.github.klee0kai.test.tech.phone.base;
+
+public interface APhoneLifecycle {
+
+    void onBuy();
+
+    void onBroken();
+
+    void onDrown();
+
+}
