+++ tests/src/test/java/com/github/klee0kai/stone/test/inject/InjectTests.java
@@ -1,9 +1,9 @@
 package com.github.klee0kai.stone.test.inject;
 
 import com.github.klee0kai.test.inject.Forest;
-import com.github.klee0kai.test.inject.Horse;
-import com.github.klee0kai.test.inject.Mowgli;
-import com.github.klee0kai.test.inject.Snake;
+import com.github.klee0kai.test.mowgli.animal.Horse;
+import com.github.klee0kai.test.mowgli.animal.Mowgli;
+import com.github.klee0kai.test.mowgli.animal.Snake;
 import org.junit.jupiter.api.Test;
 
 import java.util.Objects;
