+++ tests/src/test/java/com/github/klee0kai/stone/test/inject/ProtectInjectTests.java
@@ -1,10 +1,10 @@
 package com.github.klee0kai.stone.test.inject;
 
 import com.github.klee0kai.test.inject.Forest;
-import com.github.klee0kai.test.inject.Horse;
-import com.github.klee0kai.test.inject.School;
-import com.github.klee0kai.test.inject.forest.History;
-import com.github.klee0kai.test.inject.identity.Knowledge;
+import com.github.klee0kai.test.mowgli.animal.Horse;
+import com.github.klee0kai.test.mowgli.School;
+import com.github.klee0kai.test.mowgli.world.History;
+import com.github.klee0kai.test.mowgli.identity.Knowledge;
 import org.junit.jupiter.api.Test;
 
 import java.lang.ref.WeakReference;
