+++ tests/src/test/java/com/github/klee0kai/stone/test/lifecycle/LifecycleTests.java
@@ -1,12 +1,12 @@
 package com.github.klee0kai.stone.test.lifecycle;
 
-import com.github.klee0kai.test.lifecycle.GoodPhone;
-import com.github.klee0kai.test.lifecycle.OnePhone;
+import com.github.klee0kai.test.tech.phone.GoodPhone;
+import com.github.klee0kai.test.tech.phone.OnePhone;
 import com.github.klee0kai.test.lifecycle.di.qualifiers.DataStorageSize;
 import com.github.klee0kai.test.lifecycle.di.qualifiers.RamSize;
-import com.github.klee0kai.test.lifecycle.structure.Battery;
-import com.github.klee0kai.test.lifecycle.structure.DataStorage;
-import com.github.klee0kai.test.lifecycle.structure.Ram;
+import com.github.klee0kai.test.tech.components.Battery;
+import com.github.klee0kai.test.tech.components.DataStorage;
+import com.github.klee0kai.test.tech.components.Ram;
 import org.junit.jupiter.api.Test;
 
 import java.lang.ref.WeakReference;
