+++ tests_ext/src/main/java/com/github/klee0kai/test_ext/inject/di/DiseasesModule.java
@@ -2,7 +2,7 @@ package com.github.klee0kai.test_ext.inject.di;
 
 import com.github.klee0kai.stone.annotations.module.Module;
 import com.github.klee0kai.stone.annotations.module.Provide;
-import com.github.klee0kai.test_ext.inject.diseases.Osteoarthritis;
+import com.github.klee0kai.test_ext.inject.mowgli.diseases.Osteoarthritis;
 
 @Module
 abstract public class DiseasesModule {
