+++ tests_ext/src/main/java/com/github/klee0kai/test_ext/inject/di/OldForestComponent.java
@@ -3,7 +3,7 @@ package com.github.klee0kai.test_ext.inject.di;
 import com.github.klee0kai.stone.annotations.component.Component;
 import com.github.klee0kai.stone.types.lifecycle.IStoneLifeCycleOwner;
 import com.github.klee0kai.test.inject.di.ForestComponent;
-import com.github.klee0kai.test_ext.inject.OldHorse;
+import com.github.klee0kai.test_ext.inject.mowgli.OldHorse;
 
 @Component
 public interface OldForestComponent extends ForestComponent {
