+++ tests_ext/src/main/java/com/github/klee0kai/test_ext/inject/di/OldIdentityModule.java
@@ -4,8 +4,8 @@ package com.github.klee0kai.test_ext.inject.di;
 import com.github.klee0kai.stone.annotations.module.Module;
 import com.github.klee0kai.stone.annotations.module.Provide;
 import com.github.klee0kai.test.inject.di.IdentityModule;
-import com.github.klee0kai.test_ext.inject.identity.FamilyIdeology;
-import com.github.klee0kai.test_ext.inject.identity.OldKnowledge;
+import com.github.klee0kai.test_ext.inject.mowgli.identity.FamilyIdeology;
+import com.github.klee0kai.test_ext.inject.mowgli.identity.OldKnowledge;
 
 
 @Module
