--- tests_ext/src/main/java/com/github/klee0kai/test_ext/inject/diseases/Osteoarthritis.java
@@ -1,11 +1,0 @@
-package com.github.klee0kai.test_ext.inject.diseases;
-
-
-import java.util.UUID;
-
-public class Osteoarthritis {
-
-    public UUID uuid = UUID.randomUUID();
-
-
-}
