--- tests_ext/src/main/java/com/github/klee0kai/test_ext/inject/identity/OldKnowledge.java
@@ -1,16 +1,0 @@
-package com.github.klee0kai.test_ext.inject.identity;
-
-import com.github.klee0kai.test.inject.identity.Knowledge;
-
-public class OldKnowledge extends Knowledge {
-
-
-    public boolean doChildKnowledge() {
-        return true;
-    }
-
-    public boolean isOldKnowledge() {
-        return true;
-    }
-
-}
