+++ tests_ext/src/main/java/com/github/klee0kai/test_ext/inject/mowgli/diseases/Osteoarthritis.java
@@ -1,0 +1,11 @@
+package com.github.klee0kai.test_ext.inject.mowgli.diseases;
+
+
+import java.util.UUID;
+
+public class Osteoarthritis {
+
+    public UUID uuid = UUID.randomUUID();
+
+
+}
